"cos nie gra z plaszczyznami"

+++ src/kowalski/Soczewka.java
@@ -10,7 +10,7 @@ public class Soczewka extends Element{
 	}
 	public void paint(Graphics g, int height, int width){
 	g.setColor(Color.cyan);
-    g.drawLine((int)z+(int)h1,2*height/3+(int)h/2,(int)z+(int)h1,2*height/3-(int)h/2);
-    g.drawLine((int)z+(int)h2,2*height/3+(int)h/2,(int)z+(int)h2,2*height/3-(int)h/2);
+    g.drawLine(z+(int)h1,2*height/3+(int)h/2,z+(int)h1,2*height/3-(int)h/2);
+    g.drawLine(z+(int)h2,2*height/3+(int)h/2,z+(int)h2,2*height/3-(int)h/2);
 	}
 }
